Add comments, refactor Broadcaster
* Move EngineEvent struct definition and glutin reexport from
  broadcaster.rs to event.rs

* Register EngineEvent component with Broadcaster in Context::new()
  instead of Broadcaster::new
