+++ src/context/src/broadcaster.rs
@@ -1,62 +1,49 @@
-extern crate glutin;
 extern crate amethyst_ecs;
 
-pub use glutin::{Event, ElementState, ScanCode,
-                 VirtualKeyCode, MouseScrollDelta,
-                 TouchPhase, MouseButton, Touch};
-
-use self::amethyst_ecs::{World, Component, VecStorage, EntityBuilder,
+use self::amethyst_ecs::{World, Component, EntityBuilder,
                   Storage, Allocator, MaskedStorage, Join, Entity};
 use std::sync::RwLockReadGuard;
 
-pub struct EngineEvent {
-    pub payload: Event,
-}
-
-impl EngineEvent {
-    pub fn new(event: Event) -> EngineEvent {
-        EngineEvent {
-            payload: event,
-        }
-    }
-}
-
-impl Component for EngineEvent {
-    type Storage = VecStorage<EngineEvent>;
-}
-
+/// Allows publishing entities
 pub struct Broadcaster {
     world: World,
 }
 
 impl Broadcaster {
+    /// Create new `Broadcaster`
     pub fn new() -> Broadcaster {
         let world = World::new();
-        let mut broadcaster = Broadcaster {
+        Broadcaster {
             world: world,
-        };
-        broadcaster.register::<EngineEvent>();
-        broadcaster
+        }
     }
 
+    /// Add a custom `Component` with which
+    /// entities can be built and published
+    /// using `Broadcaster::publish()`
     pub fn register<T: Component>(&mut self) {
         self.world.register::<T>();
     }
 
+    /// Build and publish an entity,
+    /// using `EntityBuilder` syntax
     pub fn publish(&mut self) -> EntityBuilder {
         self.world.create_now()
     }
 
+    /// Return a vector containing clones of all published entities
     pub fn poll(&self) -> Vec<Entity> {
         let entities = self.world.entities();
         let _entities: Vec<Entity> = entities.iter().map(|e| e.clone()).collect();
         _entities
     }
 
+    /// Access a component storage
     pub fn read<T: Component>(&self) -> Storage<T, RwLockReadGuard<Allocator>, RwLockReadGuard<MaskedStorage<T>>> {
         self.world.read::<T>()
     }
 
+    /// Delete all published entities
     pub fn clean(&mut self) {
         let entities = self.poll();
         for entity in entities {
