+++ src/engine/src/app.rs
@@ -4,7 +4,7 @@ extern crate amethyst_context;
 use super::state::{State, StateMachine};
 use super::timing::{Duration, SteadyTime, Stopwatch};
 use self::amethyst_context::Context;
-use self::amethyst_context::broadcaster::EngineEvent;
+use self::amethyst_context::event::EngineEvent;
 use std::cell::RefCell;
 use std::rc::Rc;
